fix: Move runcards mode runner to runcard
