+++ src/qibocal/auto/runcard.py
@@ -3,49 +3,20 @@
 import os
 from dataclasses import asdict
 from pathlib import Path
-from typing import Any, NewType, Optional, Union
+from typing import Optional, Union
 
 import yaml
 from pydantic.dataclasses import dataclass
-from qibolab.qubits import QubitId, QubitPairId
+from qibolab.platform import Platform
 
-from .operation import OperationId
-
-Id = NewType("Id", str)
-"""Action identifiers type."""
-
-Targets = Union[list[QubitId], list[QubitPairId], list[tuple[QubitId, ...]]]
-"""Elements to be calibrated by a single protocol."""
+from .. import protocols
+from .execute import Executor
+from .history import History
+from .mode import ExecutionMode
+from .task import Action, Targets
 
 RUNCARD = "runcard.yml"
 """Runcard filename."""
-
-SINGLE_ACTION = "action.yml"
-
-
-@dataclass(config=dict(smart_union=True))
-class Action:
-    """Action specification in the runcard."""
-
-    id: Id
-    """Action unique identifier."""
-    operation: OperationId
-    """Operation to be performed by the executor."""
-    targets: Optional[Targets] = None
-    """Local qubits (optional)."""
-    update: bool = True
-    """Runcard update mechanism."""
-    parameters: Optional[dict[str, Any]] = None
-    """Input parameters, either values or provider reference."""
-
-    def dump(self, path: Path):
-        """Dump single action to yaml."""
-        (path / SINGLE_ACTION).write_text(yaml.safe_dump(asdict(self)))
-
-    @classmethod
-    def load(cls, path):
-        """Load action from yaml."""
-        return cls(**yaml.safe_load((path / SINGLE_ACTION).read_text(encoding="utf-8")))
 
 
 @dataclass(config=dict(smart_union=True))
@@ -78,3 +49,23 @@ class Runcard:
     def dump(self, path):
         """Dump runcard object to yaml."""
         (path / RUNCARD).write_text(yaml.safe_dump(asdict(self)))
+
+    def run(
+        self, output: Path, platform: Platform, mode: ExecutionMode, update: bool
+    ) -> History:
+        """Run runcard and dump to output."""
+        targets = self.targets if self.targets is not None else list(platform.qubits)
+        history = History.load(output)
+        update = update and self.update
+        instance = Executor(
+            history=history, platform=platform, targets=targets, update=update
+        )
+
+        for action in self.actions:
+            instance.run_protocol(
+                protocol=getattr(protocols, action.operation),
+                parameters=action,
+                mode=mode,
+            )
+            instance.history.flush(output)
+        return instance.history
